Added enum to better express find-who-is-lying
Before this commit an Option was being used to identify whether one of
the peers were lying or not. This made difficult to handle cases of
unresponsiveness or both peers lying. This commit adds an enum to
enhance expressability.

It also:

Updated docstring

Removed None field of enum PeerCheck

Renamed block to block_hash

Modified PeerCheck

Added suggestions from @JoseSK999

Added check for block received

This commit adds a check to match the block received and the expected
block hash. Preventing the peer sending a fake or adultered block.

Removed code of handling block request as it is unnecessary after rebasing with the new p2p messaging protocol

Modified logic of inflight_block to handle peer timeout

Rebase, removing the future from the inflight requests